requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
jwt==1.3.1
lxml==5.3.0
numpy==2.0.2
pandas==2.2.3
PublicDataReader==1.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
redis==5.2.0
requests==2.32.3
router==0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==14.1
xmltodict==0.14.2
